[addons][peripheral] spearate "C" parts to own file
This place the "C" parts about peripheral to a new file on "kodi/c-api/addon-instance/peripheral.h".
This done to allow also "C" language only (as base for other languages) and to confirm safe "C" ABI
between addon and Kodi.
